refactor 경로수정
