+++ src/pages/Profile/Profile.jsx
@@ -7,7 +7,8 @@ import { supabase } from '@/api/supabase';
 import { useRef, useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import toast from 'react-hot-toast';
-import { useUserRecordsCount } from '@/utils/useUserRecordsCount';
+import { useUserRecordsCount } from '@/stores/useUserRecordsCount';
+
 import { clearLocalStorage } from '@/utils/clearLocalStorage';
 import Modal from '@/components/Modal/Modal';
